Wed, Jan 3, 2024, 11:27 PM +08:00

+++ readability/readability.c
@@ -19,9 +19,9 @@ int main(void)
     int sentences = count_sentences(text);
 
     // Compute the Coleman-Liau index
-    int L = (letters / words) * 100;
-    int S = (sentences / words) * 100;
-    float index = 0.0588 * ((float) L - 0.296) * ((float) S - 15.8);
+    int L = ((float) letters / (float) words) * 100;
+    int S = ((float) sentences / (float) words) * 100;
+    int index = 0.0588 * ((float) L - 0.296) * ((float) S - 15.8);
 
     // Print the grade level
     if ((int) round(index) >= 16)
